# source/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/auth/login": {
            "post": {
                "description": "Login",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Login",
                "parameters": [
                    {
                        "description": "Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.RequestLogin"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/web.ResponseStandard"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/web.ResponseStandard"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/web.ResponseStandard"
                        }
                    }
                }
            }
        },
        "/auth/read-token": {
            "get": {
                "description": "Read Token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Read Token",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/web.ResponseStandard"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/web.ResponseStandard"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/web.ResponseStandard"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "controller.RequestLogin": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "example": "admin@bumame.com"
                },
                "password": {
                    "type": "string",
                    "example": "123456"
                }
            }
        },
        "web.ErrorResponse": {
            "type": "object",
            "properties": {
                "messages": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "web.ResponseStandard": {
            "type": "object",
            "properties": {
                "data": {},
                "errors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/web.ErrorResponse"
                    }
                },
                "fulfilled": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                },
                "pagination": {
                    "description": "Meta       any `json:\"meta\"`\nJsonApi    any `json:\"jsonapi\"`\nLinks      any `json:\"links\"`\nIncluded   any `json:\"included\"`\nSelf       any `json:\"self\"`\nRelated    any `json:\"related\"`"
                },
                "statusCode": {
                    "type": "integer"
                }
            }
        }
    }
}